compiling on the laptop

+++ Utilities/socketxx/socket++/local.h
@@ -94,7 +94,7 @@ extern "C" int shutdown (int, int); // they have forgotten this
 //extern "C" char* inet_ntoa (in_addr ina);
 
 #if !(defined (__linux__) || defined(__FreeBSD__))
-  extern "C" int gethostname (char* hostname, int len);
+  extern "C" int gethostname (char* hostname, size_t len);
 #if !(defined(__FreeBSD__) || defined(__NetBSD__) || defined(__OpenBSD__) || defined(__bsdi__) || defined(__APPLE__))
   extern char* SYS_SIGLIST [];
 #endif
